chore(Pipfile): add pywin32
I switched to windows to continue the project development,
and realized that the Jupyter Notebook didn't run properly
on the pipenv environment unless pywin32 was installed.

I included it on the Pipfile to ease the life of future
contributors who use windows.

In particular, I wasn't able to start the kernel on the
Jupyter Notebook SIR-X.ipynb.